Updated Misc and old_case_match with time series methods
Temporarily have debugTMerge.R which is a copy of the current git
version of tmerge --- currently does not appear to be operating as
documented...
Not accepting the "options" indicated. Multiple time slots in timeline
were truncated into single instances in output... Further checking
required to see if adding an event adds all instances, or just event
instances. --- INCOMPLETE --- This commit is for documentation purposes
and does not yet have a functioning time series analysis.