Added ptWrapper.R to wrap the strictly upper triangular matricies. pdist_parallel (>mem methods not tested)
Adjusted pdist_parallel to only create a single vector for the distance.
Outside of pdist_parallel the patient list should be saved. Wrapper was
created for the strictly triangular matrix to make it possible to
quickly grab the column and row you want. Note that column is always >
row! not yet testing input for validity.